+++ wok_code/wok_code/scripts/please_cwd.py
@@ -355,7 +355,6 @@ class PleaseCwd(object):
         # if is_odoo or is_pypi or please.is_repo_odoo() or please.is_repo_ocb():
         if please.package.dir_level == "module":
             sts = 0
-            # for root, dirs, files in os.walk(self.cur_path_of_pkg()):
             for root, dirs, files in os.walk(please.package.path):
                 for fn in files:
                     if (
@@ -370,42 +369,27 @@ class PleaseCwd(object):
                     sts = please.os_system(cmd, with_shell=True, rtime=True)
                     if sts:
                         break
-            # logdir = please.get_logdir()
-            # if is_odoo and pth.isdir(logdir):
-            #     last = " "
-            #     for root, dirs, files in os.walk(logdir):
-            #         for fn in files:
-            #             if re.match(r".*_\d{8}.txt$", fn) and fn[-12:] > last:
-            #                 last = fn[-12:]
-            #     for root, dirs, files in os.walk(logdir):
-            #         for fn in files:
-            #             if re.match(r".*_\d{8}.txt$", fn) and fn[-12:] != last:
-            #                 cmd = "rm " + pth.join(root, fn)
-            #                 sts = please.os_system(cmd)
-            #                 if sts:
-            #                     break
             dt_limit = datetime.strftime(datetime.now() - timedelta(7), "%Y%m%d")
             ctr_min = 1
             ctr_max = 2
             ctrs = {"*": 0}
-            for fqn in please.package.list_log_filename(all_version=True):
-                ctrs["*"] += 1
-                # if ctrs["*"] < ctr_min:
-                #     continue
-                mo = re.search(
-                    r"[0-9]{2,3}-[0-9]{4}-?[0-9]{2}-?[0-9]{2}?\+[0-9]+", fqn)
-                if not mo:
-                    continue
-                ver, fqn_dt = fqn[mo.start(): mo.end()].split("-", 1)
-                ctrs[ver] = 1 if ver not in ctrs else ctrs[ver] + 1
-                if ctrs[ver] <= ctr_min:
-                    continue
-                if ctrs[ver] <= ctr_max and fqn[mo.start(): mo.end()] >= dt_limit:
-                    continue
-                cmd = "rm " + pth.join(root, fqn)
-                sts = please.os_system(cmd)
-                if sts:
-                    break
+            if please.package.log_dir:
+                for fqn in please.package.list_log_filename(all_version=True):
+                    ctrs["*"] += 1
+                    mo = re.search(
+                        r"[0-9]{2,3}-[0-9]{4}-?[0-9]{2}-?[0-9]{2}?\+[0-9]+", fqn)
+                    if not mo:
+                        continue
+                    ver, fqn_dt = fqn[mo.start(): mo.end()].split("-", 1)
+                    ctrs[ver] = 1 if ver not in ctrs else ctrs[ver] + 1
+                    if ctrs[ver] <= ctr_min:
+                        continue
+                    if ctrs[ver] <= ctr_max and fqn[mo.start(): mo.end()] >= dt_limit:
+                        continue
+                    cmd = "rm " + pth.join(root, fqn)
+                    sts = please.os_system(cmd)
+                    if sts:
+                        break
             return sts
         return please.do_iter_action("do_clean", act_all_pypi=True, act_tools=True)
 
@@ -499,13 +483,6 @@ class PleaseCwd(object):
             srcdir = os.getcwd()
             pkgname = pth.basename(srcdir)
             sts = 0
-            # if pkgname != "tools":
-            #     fn = pth.join(pth.dirname(srcdir), "setup.py")
-            #     if pth.isfile(fn):
-            #         sts = please.os_system(
-            #             "cp %s %s" % (fn, pth.join(srcdir, "scripts", "setup.info")),
-            #             rtime=True,
-            #         )
             if not please.opt_args.no_verify:
                 sts = please.os_system("git add ../", rtime=True)
                 if sts == 0:
@@ -717,14 +694,14 @@ class PleaseCwd(object):
             sts = self.run_gen_readme([])
             if sts == 0:
                 sts = self.run_gen_readme("-I")
-            if sts == 0:
+            if sts == 0 and pth.isdir(self.docs_dir):
                 saved_pwd = os.getcwd()
                 os.chdir(self.docs_dir)
                 sts = please.os_system("make html", rtime=True)
                 os.chdir(saved_pwd)
             if sts == 0:
                 self.do_clean()
-            # sts = 0  # TODO> Remoe ?
+            # sts = 0  # TODO> Remove ?
             return sts
         elif please.is_repo_odoo() or please.is_repo_ocb():
             sts, branch = please.get_odoo_branch_from_git(try_by_fs=True)
@@ -904,19 +881,6 @@ class PleaseCwd(object):
         return 126
 
     def _do_translate_export(self, action="all"):
-        # def get_po_revision_date(pofile):
-        #     po_revision_date = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-        #     with open(pofile, "r") as fd:
-        #         for ln in fd.read().split("\n"):
-        #             if "PO-Revision-Date:" in ln:
-        #                 x = re.search(
-        #                     "[0-9]{4}-[0-9]{2}-[0-9]{2}.[0-9]{2}:[0-9]{2}", ln
-        #                 )
-        #                 if ln:
-        #                     po_revision_date = ln[x.start(): x.end()]
-        #                     break
-        #     return po_revision_date
-
         please = self.please
         if action not in ("all", "translate"):
             please.log_error(
@@ -958,7 +922,6 @@ class PleaseCwd(object):
                 if self.db_name != response[0][0]:
                     please.log_warning("Database %s does not exist!" % self.db_name)
                     return 3
-                # db_create_ts = response[0][1].replace(tzinfo=None)
                 self.connect_db()
                 query = (
                     "select state from ir_module_module where name = '%s'" % module_name
@@ -1079,12 +1042,6 @@ class PleaseCwd(object):
             pkgname = pth.basename(srcdir)
             sts = 0
             if pkgname != "tools":
-                # fn = pth.join(pth.dirname(srcdir), "setup.py")
-                # if pth.isfile(fn):
-                #     sts = please.os_system(
-                #         "cp %s %s" % (fn, pth.join(srcdir, "scripts", "setup.info")),
-                #         rtime=True,
-                #     )
                 if sts == 0:
                     sts = please.os_system(
                         "vem %s update %s" % (tgtdir, pth.dirname(srcdir)), rtime=True
